requirements.txt
Django==4.2.19
djangorestframework==3.15.2
tensorflow==2.18.0
Pillow==11.1.0
opencv-python===4.11.0.86
mediapipe===0.10.21
numpy===1.26.4
keras===3.8.0
scikit-learn===1.6.1
django-cors-headers===4.7.0